side_project_FinanceAI_BE/requirements.txt

# API server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Settings and auth
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Background works
celery>=5.3.0
redis>=5.0.0

# Web crawling
httpx>=0.25.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# AI model
google-generativeai>=0.3.0

# WebSocket
websockets>=12.0

# Additional utilities
python-multipart>=0.0.6
email-validator>=2.1.0